working on bills

+++ src/main/java/com/khan/hospital_management/model/Bill.java
@@ -3,6 +3,8 @@ package com.khan.hospital_management.model;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @EqualsAndHashCode(callSuper = true)
 @Entity
 @Data
@@ -11,11 +13,12 @@ import lombok.*;
 @Builder
 @Table(name = "bills")
 public class Bill extends BaseEntity {
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(nullable = false)
-    private Double amount;
+    private BigDecimal amount;
 
     @Enumerated(EnumType.STRING)
     private PaymentStatus status;
